+++ diofant/polys/galoistools.py
@@ -1474,7 +1474,7 @@ def gf_ddf_shoup(f, p, K):
 
     """
     n = dmp_degree_in(f, 0, 0)
-    k = int(math.ceil(math.sqrt(n//2)))
+    k = math.ceil(math.sqrt(n//2))
     b = gf_frobenius_monomial_base(f, p, K)
     h = gf_frobenius_map([K.one, K.zero], f, b, p, K)
     # U[i] = x**(p**i)
