+++ diofant/simplify/hyperexpand.py
@@ -2324,7 +2324,7 @@ def _meijergexpand(func, z0, allow_hyper=False, rewrite='default', place=None):
                 integrand = z**s
                 for b in bm:
                     if not Mod(b, 1):
-                        b = int(round(b))
+                        b = round(b)
                     integrand *= gamma(b - s)
                 for a in an:
                     integrand *= gamma(1 - a + s)
@@ -2336,7 +2336,7 @@ def _meijergexpand(func, z0, allow_hyper=False, rewrite='default', place=None):
                 # Now sum the finitely many residues:
                 # XXX This speeds up some cases - is it a good idea?
                 integrand = expand_func(integrand)
-                for r in range(int(round(lu))):
+                for r in range(round(lu)):
                     resid = residue(integrand, s, b_ + r)
                     resid = apply_operators(resid, ops, lambda f: z*f.diff(z))
                     res -= resid
